pyproject.toml
[project]
name = "calvin-and-hobbes"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "udbhav1", email = "umuthakana@gmail.com" }]
dependencies = [
    "scipy>=1.13.1",
    "scikit-image>=0.23.2",
    "pillow>=10.3.0",
    "numpy>=1.26.4",
    "pymupdf>=1.24.5",
    "pypdf>=4.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
extract = "extract:main"
area = "area:main"

# [build-system]
# requires = ["hatchling"]
# build-backend = "hatchling.build"

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.rye]
managed = true
dev-dependencies = []

# [tool.hatch.metadata]
# allow-direct-references = true

# [tool.hatch.build.targets.wheel]
# packages = ["src/calvin_and_hobbes"]
